messageBroker/pythonBroker/requirements.txt
pika==1.3.2
twilio==9.2.4
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.2